Some minor bug fixes.

+++ src/calibration.cpp
@@ -266,7 +266,7 @@ void findAllPatches(const Mat& image, Size patternSize, vector<vector<Point> >& 
 
     if (DEBUG_MODE > 1)
     {
-        printf("%s << MSERs found: %d\n", __FUNCTION__, msers.size());
+        printf("%s << MSERs found: %d\n", __FUNCTION__, (int)msers.size());
     }
 
     if (DEBUG_MODE > 0)
@@ -285,7 +285,7 @@ void findAllPatches(const Mat& image, Size patternSize, vector<vector<Point> >& 
 
     if (DEBUG_MODE > 1)
     {
-        printf("%s << Total patches found = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Total patches found = %d\n", __FUNCTION__, (int)msers.size());
     }
 }
 
@@ -324,7 +324,7 @@ void randomCulling(vector<string>& inputList, int maxSearch, vector<vector<vecto
 
     int deletionIndex = 0;
 
-    printf("%s << inputList.size() = %d / %d\n", __FUNCTION__, inputList.size(), maxSearch);
+    printf("%s << inputList.size() = %d / %d\n", __FUNCTION__, (int)inputList.size(), maxSearch);
 
     while (inputList.size() > maxSearch)
     {
@@ -618,7 +618,7 @@ void findCornerPatches(Size imageSize, Size patternSize, int mode, int *XVec, in
             printf("%s << patchCentres.at(%d) = (%f, %f)\n", __FUNCTION__, i, patchCentres.at(i).x, patchCentres.at(i).y);
         }
 
-        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, remainingPatches.size());
+        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, (int)remainingPatches.size());
     }
 
     return;
@@ -660,7 +660,7 @@ void findCornerPatches(Size imageSize, Size patternSize, int mode, int *XVec, in
 
     if (DEBUG_MODE > 3)
     {
-        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, remainingPatches.size());
+        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, (int)remainingPatches.size());
     }
 
     // For all 4 corners...
@@ -704,7 +704,7 @@ void findCornerPatches(Size imageSize, Size patternSize, int mode, int *XVec, in
             printf("%s << patchCentres.at(%d) = (%f, %f)\n", __FUNCTION__, i, patchCentres.at(i).x, patchCentres.at(i).y);
         }
 
-        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, remainingPatches.size());
+        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, (int)remainingPatches.size());
     }
 
 
@@ -912,8 +912,8 @@ void findEdgePatches(Size patternSize, int mode, int *XVec, int *YVec, vector<Po
     if (DEBUG_MODE > 3)
     {
 
-        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, patchCentres.size());
-        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, remainingPatches.size());
+        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, (int)patchCentres.size());
+        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, (int)remainingPatches.size());
 
         for (unsigned int i = 0; i < patchCentres.size(); i++)
         {
@@ -1128,8 +1128,8 @@ void findInteriorPatches(Size patternSize, int mode, int *XVec, int *YVec, vecto
     if (DEBUG_MODE > 3)
     {
 
-        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, patchCentres.size());
-        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, remainingPatches.size());
+        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, (int)patchCentres.size());
+        printf("%s << remainingPatches.size() = %d\n", __FUNCTION__, (int)remainingPatches.size());
 
         for (unsigned int i = 0; i < patchCentres.size(); i++)
         {
@@ -1204,7 +1204,7 @@ void sortPatches(Size imageSize, Size patternSize, vector<Point2f>& patchCentres
 
     if (patchCentres.size() != desiredPatchCount)
     {
-        printf("%s << ERROR. An incorrect number (%d vs %d) of patches have been allocated.\n", __FUNCTION__, patchCentres.size(), desiredPatchCount);
+        printf("%s << ERROR. An incorrect number (%d vs %d) of patches have been allocated.\n", __FUNCTION__, (int)patchCentres.size(), desiredPatchCount);
 
         //patchCentres.clear();
 
@@ -1302,7 +1302,7 @@ void reorderPatches(Size patternSize, int mode, int *XVec, int *YVec, vector<Poi
     if (DEBUG_MODE > 3)
     {
 
-        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, patchCentres.size());
+        printf("%s << patchCentres.size() = %d\n", __FUNCTION__, (int)patchCentres.size());
 
         for (unsigned int i = 0; i < patchCentres.size(); i++)
         {
@@ -1391,7 +1391,7 @@ bool findPatternCorners(const Mat& image, Size patternSize, vector<Point2f>& cor
     bool found = refinePatches(image, patternSize, msers, patchCentres2f, mode);
 
 
-    if (DEBUG_MODE > 1) printf("%s << Patches found after refinement = %d\n", __FUNCTION__, msers.size());
+    if (DEBUG_MODE > 1) printf("%s << Patches found after refinement = %d\n", __FUNCTION__, (int)msers.size());
 
     if (DEBUG_MODE > 2) debugDisplayPatches(image, msers);
 
@@ -1989,7 +1989,7 @@ bool findPatchCorners(const Mat& image, Size patternSize, Mat& homography, vecto
     if (DEBUG_MODE > 2)
     {
         Mat cornersForDisplay(cornerEstimates);
-        printf("%s << Step 3: cornerEstimates.size() = %d\n", __FUNCTION__, cornerEstimates.size());
+        printf("%s << Step 3: cornerEstimates.size() = %d\n", __FUNCTION__, (int)cornerEstimates.size());
         debugDisplayPattern(image, cvSize(patternSize.width*2, patternSize.height*2), cornersForDisplay);
         printf("%s << DONE.\n", __FUNCTION__);
     }
@@ -3188,8 +3188,8 @@ bool correctPatchCentres(const Mat& image, Size patternSize, vector<Point2f>& pa
     printf("%s << Redistorting MSER centroids...\n", __FUNCTION__);
     //patchCentres.clear();
     redistortPoints(newCentres, patchCentres, cameraMatrix, distCoeffs, newCamMat);
-    printf("%s << newCentres.size() = %d\n", __FUNCTION__, newCentres.size());
-    printf("%s << patchCentres.size() = %d\n", __FUNCTION__, patchCentres.size());
+    printf("%s << newCentres.size() = %d\n", __FUNCTION__, (int)newCentres.size());
+    printf("%s << patchCentres.size() = %d\n", __FUNCTION__, (int)patchCentres.size());
     printf("%s << DONE.\n", __FUNCTION__);
 
     if (DEBUG_MODE > 2)
@@ -3393,7 +3393,7 @@ bool findPatternCentres(const Mat& image, Size patternSize, vector<Point2f>& cen
 
     if (DEBUG_MODE > 1)
     {
-        printf("%s << Patches found after refinement = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Patches found after refinement = %d\n", __FUNCTION__, (int)msers.size());
     }
 
     if (DEBUG_MODE > 3)
@@ -4124,7 +4124,7 @@ void clusterFilter(vector<mserPatch>& patches, vector<vector<Point> >& msers, in
 
         if (DEBUG_MODE > 1)
         {
-            printf("%s << Cluster Size = %d\n", __FUNCTION__, newMsers.size());
+            printf("%s << Cluster Size = %d\n", __FUNCTION__, (int)newMsers.size());
         }
 
         // Test
@@ -4157,7 +4157,7 @@ void clusterFilter(vector<mserPatch>& patches, vector<vector<Point> >& msers, in
             clusterFound = true;
             if (DEBUG_MODE > 1)
             {
-                printf("%s << Cluster found. size = %d.\n", __FUNCTION__, newPatches.size());
+                printf("%s << Cluster found. size = %d.\n", __FUNCTION__, (int)newPatches.size());
 
             }
         }
@@ -4252,7 +4252,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
     if (DEBUG_MODE > 1)
     {
         printf("%s << totalPatches = %d\n", __FUNCTION__, totalPatches);
-        printf("%s << Patches found before refinement = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Patches found before refinement = %d\n", __FUNCTION__, (int)msers.size());
     }
 
 
@@ -4265,7 +4265,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
 
     if (DEBUG_MODE > 3)
     {
-        printf("%s << Patches found before filtering = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Patches found before filtering = %d\n", __FUNCTION__, (int)msers.size());
 
         //color = Scalar(255, 255, 0);
         image.copyTo(imCpy);
@@ -4290,7 +4290,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
     {
         t = getTickCount() - t;
         printf("%s << Shape Filter duration: %fms\n", __FUNCTION__, t*1000/getTickFrequency());
-        printf("%s << Patches found after shape filter = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Patches found after shape filter = %d\n", __FUNCTION__, (int)msers.size());
     }
 
     if (DEBUG_MODE > 6)
@@ -4347,7 +4347,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
     {
         if (DEBUG_MODE > 1)
         {
-            printf("There are an insufficient (%d/%d) number of patches after enclosure filter.\n", msers.size(), totalPatches);
+            printf("There are an insufficient (%d/%d) number of patches after enclosure filter.\n", (int)msers.size(), totalPatches);
         }
         return false;
     }
@@ -4355,7 +4355,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
 
     if (DEBUG_MODE > 1)
     {
-        printf("%s << Patches found after enclosure filter = %d\n", __FUNCTION__, msers.size());
+        printf("%s << Patches found after enclosure filter = %d\n", __FUNCTION__, (int)msers.size());
     }
 
     if (DEBUG_MODE > 6)
@@ -4400,7 +4400,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
 
         if (DEBUG_MODE > 1)
         {
-            printf("%s << Patches found after cluster filter = %d\n", __FUNCTION__, msers.size());
+            printf("%s << Patches found after cluster filter = %d\n", __FUNCTION__, (int)msers.size());
         }
 
         if (DEBUG_MODE > 6)
@@ -4420,7 +4420,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
 
         if (DEBUG_MODE > 1)
         {
-            printf("%s << Patches remaining after area-based reduction = %d\n", __FUNCTION__, msers.size());
+            printf("%s << Patches remaining after area-based reduction = %d\n", __FUNCTION__, (int)msers.size());
         }
 
         if (DEBUG_MODE > 6)
@@ -4472,7 +4472,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
 
     bool acceptable = false;
     
-    if (DEBUG_MODE > 1) printf("%s << msers.size() = (%d); totalPatches = (%d); patternSize = (%d, %d)\n", __FUNCTION__, msers.size(), totalPatches, patternSize.width, patternSize.height);
+    if (DEBUG_MODE > 1) printf("%s << msers.size() = (%d); totalPatches = (%d); patternSize = (%d, %d)\n", __FUNCTION__, (int)msers.size(), totalPatches, patternSize.width, patternSize.height);
 
     if (msers.size() == totalPatches)      // if the correct number has now been found, clear the old set and add the new set
     {
@@ -4500,7 +4500,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
     {
         if (DEBUG_MODE > 1)
         {
-            printf("%s << Too many final patches = %d/%d\n", __FUNCTION__, msers.size(), totalPatches);
+            printf("%s << Too many final patches = %d/%d\n", __FUNCTION__, (int)msers.size(), totalPatches);
         }
         return false;
     }
@@ -4508,7 +4508,7 @@ bool refinePatches(const Mat& image, Size patternSize, vector<vector<Point> >& m
     {
         if (DEBUG_MODE > 1)
         {
-            printf("%s << Too few final patches: %d/%d\n", __FUNCTION__, msers.size(), totalPatches);
+            printf("%s << Too few final patches: %d/%d\n", __FUNCTION__, (int)msers.size(), totalPatches);
         }
         return false;
     }
